ART: Implemented the Biased Lock
This patch implemented the biased lock on the top of the existing thin
lock. One thread could bias a lock in unlock state to itself via a CAS
(Compare-And-Swap) instruction. When a lock is biased to a thread, the
expensive CAS operations in acquiring the lock by the thread could be
eliminated. If there is another thread trying to acquire the lock, the
bias lock will be revoked to the thin lock.

The x86 and x86_64 port were done, the biased lock could improve the 
performance of monte_carlo in Scimark (SPECjvm2008) by around 10% on 
Intel devices.

Change-Id: I9f14bbdff6b2b3527601a5a2b8cf0a39cf88511c

Signed-off-by: Yu Li <yu.l.li@intel.com>
Signed-off-by: Haitao Feng <haitao.feng@intel.com>