ART: Implemented the Biased Lock
This patch implemented the biased lock on the top of the existing thin
lock. One thread could bias a lock in unlock state to itself via a CAS
(Compare-And-Swap) instruction. When a lock is biased to a thread, the
expensive CAS operations in acquiring the lock by the thread could be
eliminated. If there is another thread trying to acquire the lock, the
bias lock will be revoked to the thin lock.

Change-Id: I9f14bbdff6b2b3527601a5a2b8cf0a39cf88511c

Signed-off-by: Yu Li <yu.l.li@intel.com>
Signed-off-by: Haitao Feng <haitao.feng@intel.com>