ART: Implemented the Biased Lock
This patch implemented the biased lock on the top of the existing thin
lock. One thread could bias a lock in unlock state to itself via a CAS
(Compare-And-Swap) instruction. When a lock is biased to a thread, the
expensive CAS operations in acquiring the lock by the thread could be
eliminated. If there is another thread trying to acquire the lock, the
bias lock will be revoked to the thin lock.

The x86 and x86_64 port were done, the biased lock could improve the
performance of monte_carlo in Scimark (SPECjvm2008) by around 10% on
Intel devices.

Change-Id: I2912e895b574b644332055045dc7a2d128a156fb
Signed-off-by: Yu Li <yu.l.li@intel.com>
Signed-off-by: Haitao Feng <haitao.feng@intel.com>