ART: ID-matching syntactic sugar for Checker
Tests often need to match against the ID of an instruction/register.
This patch adds support for a '##' marker which, when used inside
a regex, will match any positive integer.

This patch also updates the tests to make use of this feature and
refactors construction of regular expressions inside Checker.

Change-Id: I5f854a25707e44ed2fa1673ff084990e8f43e4a2