ART: clear redundant dirty cards for sticky GC
Clear the dirty cards of alloc spaces when scanning them in pause phase
because the existing sticky GC is single generational, then the
following GC cycle doesn't need the dirty cards any more from the
previous GC cycle.

The dirty cards of immune region will not be cleared in order to keep
the references from the immune region to the alloc spaces.

Change-Id: I519f071d954a589aa33dbce0cdba405f2d2cef71
Signed-off-by: Lei Li <lei.l.li@intel.com>