[WIP] Update instrumentation stubs on non resolved classes
If we try to update method entrypoints with instrumentation stubs, we
need the class to be in the kStatusResolved state. Otherwise, we fail
a DCHECK in ArtMethod::GetMethodIndex which is used to get
the method's compiled code (from the oat file).

This situation can happen when the debugger requests an event that
will cause a full deoptimization (like a METHOD_ENTRY event) while we
are loading a new class. Because we suspend all threads to update
instrumentation stubs, we may visit a class that is being loaded but
not yet resolved.

To avoid that, we now skip non-resolved classes during the full
deoptimization (which happens when all threads are suspended) and
update its methods' entrypoints for instrumentation only when the
class becomes resolved (which happens after we resule all threads).

// TODO update instrumentation only if needed: we have to know if
instrumentation has changed while we are loading the class.

Bug: 19012386
Bug: 18766029
Change-Id: I5a645dfaf5c25dcf4282c1aaeb24f1b6333baa37