ART: Implemented the Biased Lock
This patch implemented the biased locking with profiling to reduce the
lock revocation for ART. When one object was historically locked by only
one thread for a given times, its lock will change into "biased" state
and bias to its owner thread. When a lock is biased to a thread, the
expensive CAS (Compare-And-Swap) operation in acquiring the lock by the
thread could be eliminated. If there is another thread trying to
acquiring a biased lock, the biased lock will be revoked to an
unbiasable thin lock.

The biased locking scheme improved the lock acquisition performance
for the objects that were only locked by one thread and tried to reduce
the lock revocation overhead by profiling for the objects that were
locked by more than one thread.

The biased locking could improve the performance of monte_carlo in
Scimark (SPECjvm2008) by around 10% on Intel devices.

Change-Id: I2912e895b574b644332055045dc7a2d128a156fb
Signed-off-by: Yu Li <yu.l.li@intel.com>
Signed-off-by: Haitao Feng <haitao.feng@intel.com>