Refactor register access from StackVisitor
This CL simplifies error handling (used for the debugger) when a
physical register is not accessible (read/write) in the context
from the StackVisitor.

Instead of delegating the error handling to the Context class (and
its subclasses), we move that up to the StackVisitor. We add the
IsAccessibleGPR and IsAccessibleFPR methods which respectively
indicate whether a GPR or an FPR can be read from/written to
the context. That allows to simplify the API of the Context class.

Bug: 19106446
Change-Id: I6e707608d935a71571d0e975a6e766053de3763a