WIP: art: Refactor RuntimeOptions/ParsedOptions
A (mostly done) work-in-progress to refactor the RuntimeOptions to be a
type-safe map (VariantMap, see runtime_options.h) and the ParsedOptions
to delegate the parsing to CmdlineParser (see cmdline/cmdline_parser.h).

TODO: Finish the glue in ParsedOptions::Init to return a map, and then
ParsedOptions itself can become just some static helper methods.

Change-Id: Ic67c6bca5e1f33bf2ec60e2e3ff8c366bab91563