ARM/ARM64: Make register debug names more friendly to c1visualizer.
Dump registers with type(single character)+number so that they can be
recognized as a register by c1visualizer. In this way, we can highlight
the same register in different instructions easily.

This can not be applied to x86 because x86 sometimes use characters
to present a register, such as eax.

Change-Id: Ic7f43b1c7389d2a8c0c5edae04aed74c68d78a02