RFC: Opt Compiler: Try to propagate the return register.
The register allocator now tries to propagate the return register required
by HReturn. The location of the HIr generating the return value is updated
only if it is unallocated.

Please note that this patch just scratches the surface. The more generic
problem we have to deal with is propagating fixed locations, for all the HIRs,
not just HReturn.

Change-Id: I0b84ee697bdee6ceea6803c7c9da7eddaf8154cc
Signed-off-by: Serban Constantinescu <serban.constantinescu@arm.com>