[WIP] Fix transaction aborting
The exception thrown when the transaction aborts used to indicate
there was an issue during the class initialization and we need to
rollback all the changes made during that phase. However there is
nothing preventing the code from catching that exception (like catch
"all" handlers for monitor-enter/monitor-exit). Therefore we forget
the transaction aborts and we do not rollback changes.

To fix this, we now rely on the new Transaction::aborted_ field to
know whether a transaction aborted. If the exception thrown during
abort is caught later, we throw it again right after coming back
from the <clinit> so we can rollback changes.

Bug: 19202032
Change-Id: Iedc6969528a68bbdf3123146e990df4dbc57834b