WIP: Add suspend checks to special methods.
Generate suspend checks at the beginning of special methods.
If we need to call to runtime, go to the slow path where we
create a simplified but valid frame, preserve the necessary
arguments, call art_quick_test_suspend, restore arguments
and return back to the fast path. This keeps the fast path
overhead to a minimum.

FIXME: aosp_flounder-userdebug doesn't boot.
TODO: Rewrite the hack for pruning GC maps.

Bug: 19245639
Change-Id: I3de5aee783943941322a49c4cf2c4c94411dbaa2