[optimizing compiler] Enable x86 long support
With the recent support for register pairs in ParallelMoveResolver and
the X86 long ABI change, supporting X86 longs is simpler.

Remove the check for long operations, and add some missing support for
RegisterPair to get this to pass test-art-host.

Add support for Long and Double constants.

Dies in the register allocator now.

Change-Id: I9006972a65a1f191c45691104a960366747f9d16
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>