[optimizing] Enable x86 long support
With the recent support for register pairs in ParallelMoveResolver and
the X86 long ABI change, supporting X86 longs is simpler.

Remove the check for long operations, and add some missing support for
RegisterPair to get this to pass test-art-host.

Add support for Long and Double constants.

Handle the tricky load of long long if 'low' register is the same as the
base (and/or index) register.

Change HParallelMove to allow AddMove to split register pairs into
separate moves if needed for a given architecture.  This is needed for
X86 only as the ABI uses non even/even+1 register pairs.

I added a DebugDump routine that can be called from a debugger to help
understand where things are going wrong during LSRA/Codegen.  Not sure
if this is okay with you, or you would prefer that I remove it.  I find
that debugging codegen/LSRA is very difficult without a tool like this.

Changed oatdump to dump stack mask indicies as well.  They were needed
to figure out why 103-string-append was failing.

Fixed a problem with setting the stack bit for a volatile register saved
for a call to TestSuspend that already had a stack slot.  I'm not
certain if this is the right fix, and would appreciate comments on it.

Added back in ValidateInternal calls.  Still need to split out callee
save change as requested.

Change-Id: I9006972a65a1f191c45691104a960366747f9d16
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>