[WIP][optimizing] Enable x86 long support
With the recent support for register pairs in ParallelMoveResolver and
the X86 long ABI change, supporting X86 longs is simpler.

Remove the check for long operations, and add some missing support for
RegisterPair to get this to pass test-art-host.

Add support for Long and Double constants.

Change HParallelMove to allow AddMove to split register pairs into
separate moves if needed for a given architecture.  This is needed for
X86 only as the ABI uses non even/even+1 register pairs.

Runs all the optimizing tests except for 006-args.  I still have to
track that one down.  I haven't tried BTG yet.

I added a DebugDump routine that can be called from a debugger to help
understand where things are going wrong during LSRA/Codegen.  Not sure
if this is okay with you, or you would prefer that I remove it.  I find
that debugging codegen/LSRA is very difficult without a tool like this.

Change-Id: I9006972a65a1f191c45691104a960366747f9d16
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>