[WIP] DO NOT MERGE: Wconversion
Up for a discussion. To turn on Wconversion to warn on precision
loss of implicit conversion (would have caught a bug in the x86-64
assembler), quite a number of changes are necessary. size_t is prolific,
but the mirror objects all take 32b values. Dex things are often 16b.
Etc.

This CL can compile the first 15 or so .cc files and is by far not
complete. It also turns of sign-conversion warnings, so that the
prolific mismatch of int32_t and uint32_t doesn't matter.

Change-Id: I16de682cfef336bc9c601ca47a53f2468d455283