Quick: Disable DCE until we fix FP/Core reg mismatches.
GVN and TypeInference can miss some FP/Core register
mismatch:
    const v0, #X     ; marked as float
    invoke-virtual v0, void foo(float)
    const v0, #X     ; marked as core
    aput v0, v1, v2  ; float[] not used
The second const is eliminated by the DCE because it writes
the same value to the same dalvik reg. When we generate code
for the invoke-virtual, we get a non-FP location and may
need a temporary to set up the arguments and we may not have
an available temporary.

Change-Id: I44abd390706b980c346b36c9828a3bcb4e96726d