Opt Compiler: Take into account the register type when allocating regs.
This CL extends the register allocation algorithm taking into account
the register type (callee/caller saved). If the interval we are
currently allocating a register for overlaps with a location that
will call, we prefer callee saved registers. Similarly if the interval
does not overlap with any call we prefer caller saved registers.

Change-Id: Id7ce5b3330efc33f2afb92c573eabd1a839f699f
Signed-off-by: Serban Constantinescu <serban.constantinescu@arm.com>