[MIPS] Fix mips32 trampoline problem.
Reserve argument slots (for $a0-$a3) on the stack in the
InvokeTrampoline() function, for entrypoints that contain direct
reference to native implementation. Called function may use this
space to store $a0-$a3 regs.

This fixes the Settings app crash on mips32, when navigating to:
Settings->Apps->(swipe) On SD card.

Change-Id: If53ce822fd3ef6ef9839c89b556f76fac0792190