Change thread unregister logic
There was a race where Thread::join would return before the thread was
unregistered. This caused a problem with Daemons.stop since the thread
list could get deleted before the daemon thread was removed from list_.
This caused occasional "Request to unregister unattached thread"
errors and warnings.

The fix is to remove the thread from the list before clearing the
native peer.

Bug: 18713034
Change-Id: I8921122fe8462643a6b814b5f00632481e3831fb