deopt-based bce.
A mechanism is introduced that a runtime method can be called
from code compiled with optimizing compiler to deoptimize into
interpreter. We can then use the mechanism to prove some invariants
that are expected to hold earlier in the method, and can optimize the
code further in the method with the invariants. Interpreter will do
the right thing if the invariants fail, e.g. throw exceptions according
to the spec.

The optimizatin comes with this change is for array accesses like
array[0], array[1], array[2] in a block, we add a comparison
  if (array.length < 3) deoptimize();
earlier in the block. And we can then eliminate all the three bounds checks.

Change-Id: I8c6e34b56c85d25b91074832d13dba1db0a81569