Fix dex2dex logic
There was a bug in the dex2dex logic where would only dex2dex if
IsCandidateForCompilation was true or method verifier had check casts.

The bug was caused by not adding verification results, which caused
verification_results_->GetVerifiedMethod to not have verified methods
in most cases for interpreter-only and verify-none compilation.

Bug: 17950037

Change-Id: I535ab8784b434b0a8e3f66e3ab4936d10a1b04bf