Opt compiler: Instruction simplification for HAdd, HNeg, HNot, and HSub.
Under assumptions for the 'cost' of each IR (eg. neither HAdd nor HSub are
faster than the other), transformations are only applied if they (locally)
cannot degrade the quality of the graph.  The code could be extended to look at
uses of the IRs and detect more opportunities for optimisations.
The optimisations in this patch do not look at other uses for their inputs.

Change-Id: Ib60dab007af30f43421ef5bb55db2ec32fb8fc0c