[WIP] Use optimizing for apps.
Tentative, not ready to submit yet. Just wanted some feedback whether
you think that's the right level.

Doing this way doesn't mess up with our current testing.

Another option would be to move this to PackageManager, but it really
is something internal other parts of the system should not know about.

Change-Id: I76c58f11f0d4756b894e797fc52712453e35adb1