ART: prevent patchoat from symlinking with dummy files
If OAT file is position independent (PIC, dex2oat's option
--compile-pic) then the patchoat just creates a symlink with
the input oat file. But if the input oat file is set by its
file descriptor (option --input-oat-fd) or just has been
inflated to a temporary file (options --input-oat-gz-file
or --input-oat-gz-fd) then its file name is set to a dummy
'input-oat-file' and making a symlink with this dummy path
makes no sense.

This patch fixes the patchoat to make a copy of the input
file in such cases.

Change-Id: Ib63b0c80b1da06b60cbb754206c510a7396624a9
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>